moved Samurai splitting algorithm to its own package
