+++ samurai/frequency_table_test.go
@@ -1,4 +1,4 @@
-package splitters
+package samurai
 
 import (
 	"testing"
